Better cleanup and running of containers

+++ Clean.ps1
@@ -1,18 +1,26 @@
-$MissionControl = docker ps --filter "name=mission_control" -qa
+./MiR-simulator/Stop-Unity.ps1
+
+$MissionControl = docker ps --filter name=mission_control -qa
 if ($MissionControl)
 {
-    Write-Host "Removeing old Mission control containers..."
+    Write-Host "Removing old Mission control containers..."
     docker stop $MissionControl
     docker rm $MissionControl
 }
-$MissionCreator = docker ps --filter "name=mission_creator" -qa
+$MissionCreator = docker ps --filter name=mission_creator -qa
 if ($MissionCreator)
 {
-    Write-Host "Removeing old Mission creator containers..."
+    Write-Host "Removing old Mission creator containers..."
     docker stop $MissionCreator
     docker rm $MissionCreator
 }
 
-Write-Host "Removing old agent containers"
-docker stop $(docker ps -q --filter name=agent)
-docker rm $(docker ps -aq --filter name=agent)
+if (docker ps --filter name=agent -q)
+{
+    docker stop $(docker ps -q --filter name=agent)
+}
+if (docker ps --filter name=agent -qa)
+{
+    Write-Host "Removing old agent containers"
+    docker rm $(docker ps -aq --filter name=agent)
+}
